% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_welch.R
\name{summary_welch}
\alias{summary_welch}
\title{Calculate Welch–Satterthwaite Equation Degrees of Freedom}
\usage{
summary_welch(object, correlation = FALSE, symbolic.cor = FALSE, ...)
}
\arguments{
\item{y}{A numeric vector.}

\item{x}{A binary vector.}
}
\value{
A positive number
}
\description{
Calculate Welch–Satterthwaite Equation Degrees of Freedom
}
\examples{
\dontrun{
welch_df(y = c(0,1,2,3), x = c("A","B", "A", "B"))
welch_df(y = c(0,1,2,3,9), x = c(0,1,0,1,0))
}
}
