% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welch_df.R
\name{welch_df}
\alias{welch_df}
\title{Summary Using Welch–Satterthwaite Equation Degrees of Freedom}
\usage{
welch_df(y, x)
}
\arguments{
\item{y}{A numeric vector.}

\item{x}{A binary vector.}
}
\value{
A positive number
}
\description{
Summary Using Welch–Satterthwaite Equation Degrees of Freedom
}
\examples{
\dontrun{
welch_df(y = c(0,1,2,3), x = c("A","B", "A", "B"))
welch_df(y = c(0,1,2,3,9), x = c(0,1,0,1,0))
}
}
